Discards Ruby on Rails as backend
It is very complex and hacky to use pure Websockets on Rails. I discard using Action Cable because it requires some middleware on the client side.

Also this method would not take advantage from any benefit from using a structured MVC framework such as Rails.

More info from the author of Iodine: https://bowild.wordpress.com/2016/07/31/the-dark-side-of-the-rack/